requirements.txt
cerebras-cloud-sdk
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
email-validator==2.0.0
schedule==1.2.0